improved room separation line reporting
